refactor: move db emoji handling to separate file
